// HealthSysHub.Web.Pharmacy.UI/src/app/models/profile-update-request.ts
export interface ProfileUpdateRequest {
  id?: string | null;
  entityType?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  phone?: string | null;
  hospitalId?: string | null;
  pharmacyId?: string | null;
  staffId?: string | null;
  labId?: string | null;
  modifiedBy?: string | null;
  modifiedOn?: string | null;
}